reset keep alive timer only on incoming ACK packet
keep alive timer was reset every time a packet was sent. This might be OK for the broker side but since we currently only support QoS 0, this prevented the client to catch unavailable Broker. By resetting the keep alive only on incoming ACK packet we ensure sending a PINGREQ more frequently and catch the unavailable Broker on missing PINGRESP.